Update
Initial update, I tried to get as much put into the initial mobile view as possible, will need to add a mask to the img with the accent color as well as setting the font type and adjusting font weights and sizing. Then setting up a media query for desktop.