Update
I have the mobile layout where I like it, need to set up font style and double check margin and padding. As for desktop, I have set up the media query to kick in at 1440px, gave the non image elements their own container and set it's width to 50%, as is the img container. I then adjusted the h1 font size and width to reflect the design, same with p and the stats.